Removed unused code.
