--- ui/src/main/webapp/containers/exceptions.less
@@ -1,22 +1,0 @@
-.exception {
-  position: relative;
-  background: white;
-}
-
-.message {
-  font-family: monospace;
-  font-size: 0.75em;
-  white-space: nowrap;
-  max-height: 320px;
-  overflow: scroll;
-  padding: 20px;
-  box-sizing: border-box;
-
-  a {
-    color: #333333;
-    margin-left: 4em;
-  }
-  a:hover {
-    text-decoration: underline;
-  }
-}
