+++ ui/src/main/webapp/lib/admin-wizard/actions.js
@@ -7,3 +7,5 @@ export const editConfig = (id, value) => ({ type: 'EDIT_CONFIG', id, value })
 export const editConfigs = (values) => ({ type: 'EDIT_CONFIGS', values })
 export const setDefaults = (values) => ({ type: 'SET_DEFAULTS', values })
 export const clearWizard = () => ({ type: 'CLEAR_WIZARD' })
+export const start = (id) => ({ type: 'fetch/START', id })
+export const end = (id) => ({ type: 'fetch/END', id })
